Add support to EnvironmentOrder to controller
Track the _current_ versions of channels for all active clusters,
grouped by environment and channel. If a cluster's channel version
changes, and we have an environment configured as a dependency, the
update is skipped unless all the clusters in the preceding environment
use the new version. Note that this takes precedence even over the
"already updating" case, otherwise we might get updates that shouldn't
happen in rare cases.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>