Add support for environment ordering (#15)
* Controller: move version resolution to refresh()

refresh() now gets the information about available channels and the
corresponding versions (e.g. git revision) and stores the resolved
version in the cluster list entry. All other methods, such as doProcess,
work with a specific revision instead of checking out the channel and
determining the revision when they're called. This allows us to make
sure that all updates between two refresh() calls will use the same
revision numbers for the same channel, and the revision for a particular
channel will only change by refresh().

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Move Version() to Cluster; add proper tests

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Use explicit return in doProcessCluster

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Don't access cluster state until next refresh

Change clusterInfo.processing from a boolean to an enum of {idle,
processing, processed}. Workers move the cluster from idle to
processing to processed, but only refresh() moves them back and updates
at the same time. This makes it easier to reason about concurrent access
to the cluster fields, as we can be sure that we won't read something
that we wrote with no synchronization.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Move update decision to cluster_list

ClusterList now maintains a list of clusters that need to be updated and
automatically updates it on every UpdateAvailable call. The workers now
only get a cluster from SelectNext() if something actually needs to be
updated (e.g. config items change or there's a channel update) and don't
have to constantly poll, consuming CPU on git pulls, clones, AWS calls
and other useless stuff.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Add support to EnvironmentOrder to controller

Track the _current_ versions of channels for all active clusters,
grouped by environment and channel. If a cluster's channel version
changes, and we have an environment configured as a dependency, the
update is skipped unless all the clusters in the preceding environment
use the new version. Note that this takes precedence even over the
"already updating" case, otherwise we might get updates that shouldn't
happen in rare cases.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Change profile names

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Remove unused method

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Add docs for api/cluster_version.go

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* NewStaticVersions: return a concrete type

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Document orderByEnvironmentOrder

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>