ClusterList/Controller: handle blocked updates
Add support for a cluster-update-block config item. If present, the
cluster will not be selected for update until it's removed.
Additionally, SelectNext() now accepts a context.CancelFunc which will
be automatically called on the next refresh if the cluster updates are
blocked.

Controller: create a context on every update and pass it to provisioner.
This is ignored for now, real support will be implemented in follow-up
commits.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>