Read configuration defaults from a file
Check if cluster/config-defaults.yaml exists as the first step during
provisioning. If it's there, evaluate it as a Go template with a cluster
definition stripped of config items, parse it as a YAML file that's
expected to be a simple string->string map, and set all missing config
items to the values from the parsed map.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>