Buffer pods: use pod count when calculating the size
Update autoscalingBufferSettings to pay attention to
autoscaling_buffer_pods as well. If it's 0, return a dummy 100m/100Mi
size immediately without checking other settings, otherwise divide the
buffer size by the expected number of replicas. This makes the
evictions a bit more fine-grained (so not the whole node is vacated at
once) and side-steps the issue of having to update the settings every
time we change the daemonsets.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>