Unify template data, add proper AZ support
Instead of ad-hoc data objects passed to each template, pass one object
that so far includes everything (for migration purposes). Some of the
fields might not be available (e.g. NodePool will be missing for most of
the manifests), but at least the rest will always be there.

Add support for availability_zones property.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>