use in-memory instance data
This refactors the code to pass on the instance data fetched in the
Provision function of CLM. The renderTemplate has been extended to
include the InstanceTypes type which is then passed to other functions
like renderSingleTemplate.

prepareProvision function has been extended to return the instanceTypes
data which is used in Provision function. This makes sure that we only
fetch the data from AWS API once and use it everywhere.

Additionally, NodePoolTemplateRenderer has been extended to include
instanceTypes information centrally and that has been removed from the
AWSNodePoolProvisioner.

Signed-off-by: Noor Malik <noor.malik@zalando.de>